+++ corehq/apps/app_manager/static/app_manager/js/forms/advanced/actions.js
